Fix issue with test_syslog_rate_limit (#10867)
There is a race condition in the rate-limit-host
Changing the rate limit via the cli triggers hostcfgd to restart rsyslogd.
However there is no wait to make sure this actually happens.
This leads to some flakyness where loganalyzer.py would try to put a marker while
the rsyslog systemd service is being restarted and crash.

Signed-off-by: yenlu <yenlu@cisco.com>