Fix test_counterpoll_watermark flakiness (#10877)
Description of PR
This test randomly picks a counter to test out of 3.
One check is only meant to run on one of these counter but ends up running for all counters.
There is another call for this check that is properly gated a few lines above, this fix essentially make sure it works the same.

The check actually works for all counters per my investigation but the delay is not sufficient for some older products. These take longer to properly populate data in COUNTERS_DB.
For this purpose a wait_until was introduced instead of an arbitrary sleep value.

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 Test case(new/improvement)
Back port request
 201911
 202012
 202205
 202305
Approach
What is the motivation for this PR?
Reduce test flakiness to increase passrate reliability

How did you do it?
Investigated flaky test, reproduced and looked at the code to understand what went wrong

How did you verify/test it?
Ran the flaky test in a loop overnight.

Any platform specific information?
This test seems flaky on slower products

Signed-off-by: yenlu <yenlu@cisco.com>