Fix test_show_platform_ssdhealth test for 7050QX-32S (#10967)
What is the motivation for this PR?
For some reason this test was marked XFAIL though it passes fine on 202305+ It is likely related to enhancements made both to ssdutil and the arista platform library to gracefully handle emmc/eUSB devices

How did you do it?
Remove the pytest mark forcing XFAIL on this test for the 7050QX-32S product

How did you verify/test it?
Ran the test and check it passed reliably

Any platform specific information?
Only affecting Arista 7050QX-32S

Signed-off-by: yenlu <yenlu@cisco.com>