Merged PR 4169: Update the 202012 -> 202305 warm upgrade pipelines
For 7050cx3 (on vms28-t0-7050-14), switch to using the latest 202305 image instead of my old private image. The latest image should have successful warm upgrades at this point, with all necessary SAI fixes in.

For 7260 (on vms7-t0-7260-1), since nightly tests will not run on this testbed, change the schedule to running every day in both the morning and evening (midnight and noon UTC time), and run 18 iterations of warm upgrades.

For 7060 (on vms20-t0-7060), add a pipeline definition that would allow running 202012 -> 202305, but don't define a schedule for it yet. The current LAG downtime appears to be around 100 seconds, which means almost every run would probably fail on the first attempt.