Allow fewer PORT_READY messages than expected (#10894)
* Allow fewer PORT_READY messages than expected

On some platforms, it seems that there are no PORT_READY messages
generated with state changes (i.e. where the port changes from down to
up) during fast-reboot. For now, loosen the check for the number of
PORT_READY messages to accept fewer messages.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

* Modify the check to allow either 0 or the expected number of PORT_READY messages

This tightens up the check by allowing only two values (or one, in the
case of warm-reboot) for PORT_READY.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>