Reorder eventd heartbeat testcase as first testcase (#10952)
Summary:
Fixes # (issue) ADO 25860806

- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] Test case(new/improvement)

I am changing the order of the testcases in test_events because eventd heartbeat testcase should be used as a preliminary test before all testcases since it is testing that a critical part of eventd which is heartbeat is working correctly. It is also good that it is executed first because we can use it to drain the eventd cache in case there were any published events missed due to cache overflow.

Code change

Manual testing/pipeline