Reorder eventd heartbeat testcase as first testcase (#10952)
## Description of PR
Summary:
Fixes # (issue) ADO 25860806

### Type of change

- [x] Bug fix
- [ ] Testbed and Framework(new/improvement)
- [ ] Test case(new/improvement)

## Approach
#### What is the motivation for this PR?

I am changing the order of the testcases in test_events because eventd heartbeat testcase should be used as a preliminary test before all testcases since it is testing that a critical part of eventd which is heartbeat is working correctly. It is also good that it is executed first because we can use it to drain the eventd cache in case there were any published events missed due to cache overflow.

#### How did you do it?

Code change

#### How did you verify/test it?

Manual testing/pipeline

Signed-off-by: yenlu <yenlu@cisco.com>