Update the pfcwd warm reboot test for "create_only_config_db_buffers" (#10944)
Summary:
The test_pfcwd_warm_reboot.py no_storm and async_storm test cases sometimes fails on MSFT SKU setups due to the change in PR:

[buffers] Add handler for the 'create_only_config_db_buffers' configuration knob sonic-swss#2883

[buffers] Add 'create_only_config_db_buffers.json' file for the Mellanox devices (not MSFT SKU) sonic-buildimage#16233.

With this change, all buffer counters (which are read from SAI) will be created after warm reboot as soon as possible, regardless of the CONFIG_DB buffers configuration. It will take much longer time to complete the counter creation
And the pfcwd is only available after all the counters are created, when the test sends pfc storm after warm-reboot, the counters could have not been completely created, so no storm is detected by the pfcwd. This causes the test to fail on the platforms with MSFT SKUs.
We need to add some delay after warm-reboot if "create_only_config_db_buffers" is not enabled.