Merged PR 4260: Enable 202305 upgrade pipeline on vms20-t0-7060 and vms28-t0-7050-12
Enable 202012 -> 202305 upgrade pipeline on vms20-t0-7060 and vms28-t0-7050-12. On both of these devices, the upgrade is done with vSONiC neighbors; this is necessary for 7060 currently, but optional on 7050cx3. This has the advantage of testing the teamd retry count feature on actual hardware.

Also, for the other upgrade pipeline files, define the `VM_TYPE` and `VSONIC_NEIGHBOR_IMAGE_URL` parameters. They won't really be used, and will still be using ceos neighbors, but it's being added for completeness/consistency's sake.