Send gratuitous ARP packet when bringing up SONiC VM
When a new SONiC VM comes up, but the IP address assigned to it was used
by something else before, there may be traffic issues until the ARP
entry for that IP address expires on the router/switch connected to the
management port. In some cases, there may be routers/switches that do
not properly send ARP requests for whatever reason, when that IP address
expires.

To fix this issue, send a gratuitous ARP reply packet after the
management IP address is configred in sonic_kickstart.py. This makes
sure traffic is working, at least.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>