Include the hostname when generating the post-reboot analysis
After doing a warm/fast reboot, in the post-reboot analysis, include the
hostname of the device. This information will then get uploaded into
Kusto (if enabled). The reason for this is that in case there are some
device-specific oddities that cause some things to take longer on one
device compared to another device, then this will help reveal them. It
also makes it easier to see what device a certain run was done on, in
case there need to be some later debugging effort.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>