[AdvancedReboot]: Split out device state checks into a separate function
Currently, the AdvancedReboot class's tearDown() function has both
checks to verify that the device is in a good state, as well as cleanup
stuff (such as going back to a previous image). If multiple warm
upgrades should be run within a test case, it would be nice if it was
possible to verify the device state, but not necessarily jump to another
image.

For this purpose, split out the device checks into a separate function,
so that they can be called separately. The tear down function will still
do the device checks, since they appear to be idempotent.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>