[chassis][voq]Fix test_fabric_reach.py to run on t2 min topology. (#10727)
Fix test_fabric_reach.py to run on t2 min topology.

With the t2 min topology, only 2 linecards get initialized. Consequently, in the current code, the links on the supervisor connected to the non-initialized cards could fail the test. This change addresses the issue, and the code will now skip checking the links on the supervisor connected to those uninitialized cards.

manually run the test on a t2 min topo, and it passed.