[chassis][voq]add jr2/j2c+ Arista hwsku (#10720)
Add all the arista DNX Linecards skus to the vars file
What is the motivation for this PR?
Currently some of the Arista voq linecards are not present in the variable files

How did you do it?
Add the missing sku names

How did you verify/test it?
run chassis tests

Signed-off-by: yenlu <yenlu@cisco.com>