Support recovering Nexus devices via console (#11264)
What is the motivation for this PR?
In PR #10806, we support auto recovery of testbeds via console. But at that time, we didn't support nexus testbeds.
In Nexus testbeds, they first enter into loader, and after rebooting in loader, they enter into ONIE. So in this PR, we enhance the function which will do recover in ONIE to cover this sistuation.

How did you do it?
Enhance the function which will do recover in ONIE to cover Nexus testbeds. If the type of testbeds is Nexus, first do reboot, and then, recover in ONIE.

Signed-off-by: yenlu <yenlu@cisco.com>