Add ip information into `/etc/network/interface` after testbed is recovered via console.  (#11318)
What is the motivation for this PR?
In PR #10806 , we support auto recovery of testbeds via console. After power cycle, we will add ip and route information on device. And in this PR, we will add these information into file /etc/network/interface to avoid losing management ip again after reboot.

How did you do it?
Give the template of /etc/network/interface and after power cycle, fill in this template and set it on device.

How did you verify/test it?
When the scripts finish running, reboot the device, and it will not lose management ip.