Merged PR 4280: Initial files for setting up TK5 lab
This change added the required files for new TK5 lab.

Currently, the server and PTF IP in this PR are not final.

This PR only covered the initial trial move of
* 3 fanout
* 1 root fanout
* 2 Cisco 8101 DUT
* 5 servers

The subnet allocated for these devices has prefix length 27. This subnet is not enough for the devices we
moved on 12/11. We are requesting new dedicated subnet with prefix length 24 for all the lab servers and PTF containers. After the new allocation is done, the servers will be moved to the new subnet. Then the server IP and PTF IP must be changed accordingly.