[Fdb/test_fdb]: Fixed fdb test case failure due to incomplete mac learning on all memebers of portchannel (#11170)
What is the motivation for this PR?
FDB test case test_fdb failed intermittantly due to incomplete mac learning on the fanout. During the test runtime, the test sends a series of packets so that the correct mac's are learnt on the interfaces and PortChannels being tested. However, for the port channels, MAC's are learnt only on the first member of the PortChannel and not all members of the PortChannel. Therefore, while testing if forwarding is correctly happening through the PortChannel, if a test packet is forwarded through the second member of the portchannel, then the testcase fails since the fanout does not know where to forward the incoming packet. The testcase however passes if the packet is forwarded through the first member of the PortChannel.

How did you do it?
The change in the PR involves including the mac learning to happen across all the members of the PortChannel. Once mac learning is enabled on all members of the PortChannel, then the testcase passes and doesnt fail intermittently.

How did you verify/test it?
Ran a script to run the test continuously and checked that it passes.