Support recovering Nokia devices via console (#11329)
What is the motivation for this PR?
In PR #10806, we support auto recovery of testbeds via console. But at that time, we didn't support nokia testbeds.
In Nokia testbeds, they first enter into u-boot, and we can enter into ONIE through u-boot. So in this PR, we enhance the function which will do recover in ONIE to cover this sistuation.

How did you do it?
Enhance the function which will do recover in ONIE to cover Nokia testbeds. If the type of testbeds is Nokia, enter into ONIE through u-boot, and then, recover in ONIE.

How did you verify/test it?
Recover Nokia testbeds.