Remove termios setting while recovering testbeds (#11362)
Approach
What is the motivation for this PR?
Termios is a python lib to set something about terminal which should be used in terminal. While doing recovering, Elastictest will not directly using a terminal, which will generate an Exception (25, 'Inappropriate ioctl for device'). Here, we don't need to set terminal, so in this PR, I remove the related code to avoid exception.

How did you do it?
Remove related code to set terminal.

How did you verify/test it?
Use Elastictest to run a testplan, and the Exception will not appear again.

co-authorized by: jianquanye@microsoft.com