[chassis][voq]Add mgmt test for testing fabric link isolation cli commands. (#10994)
Summary: Add mgmt test for testing fabric link isolation cli commands.

This change adds a test for the fabric link isolation command.
The test iterates over the fabric links on a linecard. It isolates and unisolates fabric links. Each time the state of the link is changed the state is checked in both CONFIG_DB and APPL_DB. The values in CONFIG_DB are updated by the fabric CLI commands and the values in APPL_DB are updated by the fabric manager daemon. The test check if the 2 major part work as expected.

need to wait sonic-net/sonic-buildimage#17473

Signed-off-by: Jie Feng <jfeng@arista.com>