Relax the check on the number of PORT_READY messages
Sometimes, after fast reboot, there is a random number of PORT_READY
messages that occur; it could be 0, it could be the number of ports that
are up, it could be some number in between. The reasoning for this is a
bit unclear to me. However, considering the logic here prior to my
changes had a relaxed check here, restore that behavior.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>