Fix qos sai issue
The speed support list for onyx is like '1G 10G 25G 40G 50Gx1 50Gx2 100Gx2 100Gx4 200Gx4'. When the fanout is onyx, we need to set the speed according to the speed support list. For example, when dut and fanout all support 50G, if support speed list of fanout just includes 50Gx1 not includes 50G, we need to set the speed with 50Gx1 instead of 50G, otherwise, the port can not be up.

Change-Id: I5d14151422fc10ee23fd410d999278cff59ff52b