Merged PR 4633: Save management ip into config_db.json after recovery via console.
After recovery from console and get ssh session back, we need to upgrade image. But we notice that, during upgradation, while reboot, we will lose management ip again and casue the failure. In this PR, we write management ip into config_db.json to avoid the loss of management ip.