[Github] Delete branch after "automerge" PRs are merged by workflow
We are using Github workflows to automatically create PRs for cherry-picking.
These auto created PRs are labeled with "automerge". Another workflow
periodically scans the PRs with "automerge" and merge them if all checks passed.

The problem is that the branch of these auto created PRs are not deleted after
the PRs are merged. Consequently, the mssonicbld/sonic-mgmt repo has too
 many branches left over now. We'll manually do some cleanup.

 Meanwhile, we also need to delete corresponding branch after PR is auto merged.

This changed added option "--delete-branch" to the Github cli for merging PRs.
With this change,

Signed-off-by: Xin Wang <xiwang5@microsoft.com>