[Snappi] Update pause frame counting logic for DUT (#11400)
When verifying pause frames being generated/transmitted and received by the DUT during PFC tests, the way in which received pause frames are counted for lossy test cases or when the pause is not expected to be paused is incorrect. All PFC tests receive PFC frames from the traffic generator to congest the DUT's egress port, hence the check to ensure if this number is 0 is incorrect.

Signed-off-by: yenlu <yenlu@cisco.com>