Fix iface_namingmode alias check (#10892)
Description of PR
The iface_namingmode test attempts to verify that the original interface name is not present in counters output when alias mode is requested, but it winds up matching if the original interface name is a substring of an alias name (e.g., Ethernet1 would match Ethernet12 in the output and cause the test to fail).

Approach
What is the motivation for this PR?
This PR fixes a bug where the test mistakenly matches a substring of an interface name when looking for the original name of another interface (e.g., Ethernet12 matches Ethernet1)

How did you do it?
I isolated the interface names from the CLI output and matched against the whole interface name, instead of doing a substring match on the entire output.

How did you verify/test it?
I ran the test manually against a DUT which was affected.

Signed-off-by: yenlu <yenlu@cisco.com>