[Fix]:PFC multidut helper fix for latest traffic_generator.py (#10926)
What is the motivation for this PR?
This PR is to accommodate the latest changes that went into the traffic_generation.py file

How did you do it?
Instead of passing the parameters as arguments for the fixtures that were used for creating traffic flows, the parameters were defined on snappi_extra_params object and there by reducing the arguments on the flow creation fixtures, this made the code significantly compact and scalable

How did you verify/test it?
After the changes the scripts were rerun on Arista DUT and ensured the cases run fine

Signed-off-by: yenlu <yenlu@cisco.com>