Porting the PR:9655 to 202205. (#11020)
Porting PR: #9655 to 202205:
Addressed in this PR:

Cisco-8000/Q100 platform doesn't add the headroom drops to the rx-drop counters (as this are not per port), so that particular check is being skipped in the PFCXoff testcase.
The print messages encounter this problem: formatting problem in tests/saitests/py3/qos_sai_tests.py #9652. This PR addresses the formatting.