Merged PR 4668: [Nightly] Set MIN_WORKER=1 for MGFX nightly testplans with multiple testbeds
Set MIN_WORKER=1 for MGFX nightly testplans with multiple testbeds.

Previously we set MIN_WORKER=2 because we run nightly everyday and a single TB cannot finish in 24 hours. Now ElasticTest supports replenish which can kick out the unhealthy TB and put it back to testplan once recovered. This is a great feature and we can change MIN_WORKER to 1.