Merged PR 4667: Skip activating python3 venv for non-critical pipelines
Skip activating python3 venv for non-critical pipelines

All of the packages installed in the python3 virtual environment have been
installed in the global python3 environment. The python3 virtual environment
is no longer needed and can be deprecated.

Before deprecating the python3 virtual environment, we'll need to update
all the pipelines depending on the python3 venv to use the global python3
environment.

This change firstly updated the non-critical pipelines to skip activating
python3 venv. If it works fine. We'll continue updating the critical pipelines
for nightly tests.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>