Change into the /mnt/flash directory while installing image in aboot. (#11621)
Approach
What is the motivation for this PR?
When recovering testbeds whose hwsku are arista, we use A-boot to re-install image. In our previously code, we download image in ~ directory and the set-up file boot-config is in the directory /mnt/flash, which will cause error flash:{{image_name}} not found or not a file. So in this PR, we change into the /mnt/flash directory first and download image.

How did you do it?
Change directory in A-boot first.

co-authorized by: jianquanye@microsoft.com