[Recover via console] Support recovering Celestica devices via console (#11604)
What is the motivation for this PR?
In PR #10806, we support auto recovery of testbeds via console. But at that time, we didn't support Celestica testbeds at that time. Celestica testbeds use ONIE so we directly add Celestica into the branch which will be recoverd by ONIE.

How did you do it?
Add Celestica into the branch which will be recoverd by ONIE