[Recover via console] Remove image to avoid "File exists" error (#11647)
What is the motivation for this PR?
When recovering testbeds whose hwsku are Arista, we need to re-install image in Aboot. But the previous downloaded image will not be deleted in Aboot, and if we download the same image again, it will case File exists error in Aboot. We can either remove or rename the existing image on device, but rename image may cause error free memory or disk space size is not enough. So in this PR, we will remove the same image in Aboot.

How did you do it?
Remove the same image if existing before download new image.

Signed-off-by: yenlu <yenlu@cisco.com>