Merged PR 4697: Fix the default management gw on neighbors
Fix the default management gw on neighbors

For PTF container to communicate with neighbos like cEOS,
the default route on cEOS must be pointed to the 172.x.x.x
IP address assigned to the "br1" bridge which has management
interfaces of PTF and neighbors attached.

Then the PTF and neighbors can communicate via the "br1" bridge.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>