[chassis] Update J2 100_120k qos profile to match J2c (#11560)
The J2c+ qos params for 100_120K was updated in #10146 but it wasn't updated for J2.
This has caused failures on testQosSaiPfcXoffLimit because we aren't sending sufficient packets to fill up the headroom so packets aren't being dropped.

I confirmed on a J2 system with these new values we end up filling up the headroom and drop packets.

Signed-off-by: yenlu <yenlu@cisco.com>