Extend waiting time (#11711)
Approach
What is the motivation for this PR?
While recovering via console, we need to re-install sonic image in boot loader. When installing in Aboot, we need to download the image file from a HTTP or TFTP server, and then use the boot command to load and boot the image. The installation process may take some time depending on the network speed and the image size. Previous waiting time is not enough for some situation, so in this PR, I add the waiting time while downloading image.

How did you do it?
Add the waiting time while downloading image in Aboot.

co-authorized by: jianquanye@microsoft.com

Signed-off-by: yenlu <yenlu@cisco.com>