Fit for celestica (#11743)
Approach
What is the motivation for this PR?
ONIE Service Discovery" as the symbol. However, in the Celestica scenario, we cannot catch the whole word "Discovery," no matter how we adjust the buffer we set. Therefore, in this PR, we shorten the symbol to fit the Celestica scenario.

How did you do it?
Shorten the symbol to fit the Celestica scenario.

Signed-off-by: yenlu <yenlu@cisco.com>