Merged PR 4742: Don't fail postupgrade script when running on non-SSD 7050cx3
There's been a recent addition to the postupgrade script to check the storage and memory of each device, to make sure it matches what we expect. However, there are some lab devices that are running a version of 7050cx3 that does not have a SSD and also has a different amount of memory.

For our purposes, if we're running the postupgrade script on such a device, ignore failures here.

Test pipeline run: https://dev.azure.com/mssonic/internal/_build/results?buildId=481060&view=results