[Recover via console] Change the method of downloading sonic image in ONIE. (#11768)
Approach
What is the motivation for this PR?
We use onie-nos-install to get sonic image from image server, download it and install it previously. But in bjw, it will cost some time to connect to the image server which cause timeout error in onie-nos-install itself. To avoid this error, we change the method of downloading sonic image in ONIE. First, we will use wget to get sonic image from image server and download it, which will not result in timeout issue. And then, we will use onie-nos-install to just install the image on device.

How did you do it?
First, we will use wget to get sonic image from image server and download it, which will not result in timeout issue. And then, we will use onie-nos-install to just install the image on device.

Signed-off-by: yenlu <yenlu@cisco.com>