[Chassis] Mutli_asic check for T2_arppopulate qos_test_corrected (#11706)
Description of PR
The check - if the chassis is multi_asic , should be verified with the num_asic count on the dut.
Adding the correct method to verify multi_asic chassis in sai_qos_tests
In continuation to PR #11282

Approach
For verifying the chassis is multi_asic ,the num_asic count on the dut should be greater than 1.

What is the motivation for this PR?
Found issue -
#11236

How did you do it?
How did you verify/test it?
Executed qos tests on t2 chassis & verified.

Signed-off-by: yenlu <yenlu@cisco.com>