[devutils] Fix issue of parsing output_watts of PDU
The devutils may raise exception like below if the output_watts
of PDU is empty string `''`:

ValueError("invalid literal for int() with base 10: \'\'")

This change improved the code of parsing output_watts.
Another issue is that a temporary variable named `status`
conflicts with another variable. This change also fixed that.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>