platform: daemon: chassisd: Sort keys in data_dict
The list of keys is taken unsorted from redis and is thus subject to
potential re-ordering based on how the keys are inserted during
chassisd startup. Ultimately it should not matter what order the keys
are in as long as all are present, so sort the list so that we can
compare them correctly.