[Recover via console] Avoid entering into rescue model of ONIE (#11826)
What is the motivation for this PR?
We observed a misalignment issue on 4600c dualtor testbeds, where the first line of each screen was not consistent. When we tried to select ONIE mode, it would go into ONIE Rescue mode with our previous code. To fix this, we added a condition to check if the output string contained "Rescue", and if so, we would send the arrow key "UP" to enter into ONIE Install mode.

How did you do it?
To fix this, we added a condition to check if the output string contained "Rescue", and if so, we would send the arrow key "UP" to enter into ONIE Install mode.

Signed-off-by: yenlu <yenlu@cisco.com>