Packets going to incorrect ToR fixed.
The traffic from active-active mux ports is ECMPed twice:first time on the NiC to choose the ToR, second time on the ToR to choose the uplinks. The NiC ECMP is not within the test scope, and we also cannot guarantee that the traffic is evenly distributed among all the uplinks.

The proposed fix is to configure active-active ports to work in active-standby mode in case of the following tests :-

    1. dhcp_relay/test_dhcp_relay.py
    2. dhcp_relay/test_dhcpv6_relay.py
    3. drop_packets/test_configurable_drop_counters.py
    4. pfcwd/test_pfcwd_function.py
    5. route/test_route_flap.py
    6. snmp/test_snmp_fdb.py
    7. everflow/test_everflow_testbed.py
    8. everflow/test_everflow_ipv6.py