advanced-reboot: Fix timeout not being present in Python 2
Because of #10960 needing to be cherry-picked to 202305, there was some
Python 3-only syntax present that caused failures on 202305, since the
advanced-reboot script is run on the PTF container under Python 2.
Specifically, `subprocess.wait` doesn't recognize the `timeout`
argument.

Work around that by starting a timer where a kill signal is sent to the
process if it hasn't terminated in 5 seconds. This probably won't be
actually needed, though.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>