Merged PR 4922: Update graph and inventory files for lab subnet change for 10.3.147.x
This PR is for this change from starlab:

As part of the starlab midnight blizzard remediation, we will be splitting 103.146.0/23 into two /24 subnets. 10.3.146.0/24 will be for sonic and 10.3.147.0/24 will be for non-sonic teams. Please be aware that we will make some changes in our network that will lose outside connectivity to 10.3.147.0 as they will need to be re-configured to use a different GW, and 146 will need its subnetmask changed to avoid conflicts. We are planning on making this changes on  March 8th 2024.

Simply tested this change by generating conn_graph_facts.